DSEGOG-67 moving hook to its own file, refactoring more components
Addressing PR comment by moving a hook to click outside of a dialog box to its own file

Moving components and files concerning more options in the plot settings to their own directory
