+++ src/plotting/plotSettings/moreOptions/colourPicker.component.test.tsx
